% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovmodel.R
\name{aic.markovmodel}
\alias{aic.markovmodel}
\title{Akaike Information Criterion (AIC)}
\usage{
\method{aic}{markovmodel}(x, seq, E)
}
\arguments{
\item{x}{An object of class \link{markovmodel}.}

\item{seq}{A list of vectors representing the sequences for which the
AIC criterion must be computed.}

\item{E}{Vector of state space (of length S).}
}
\value{
A numeric value giving the value of the AIC.
}
\description{
Computation of the Akaike Information Criterion.
}
