% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic.R
\name{aic}
\alias{aic}
\title{Akaike Information Criterion (AIC)}
\usage{
aic(x, sequences, states)
}
\arguments{
\item{x}{An object for which the logikelihood can be computed.}

\item{sequences}{A list of vectors representing the sequences for which the
AIC criterion must be computed.}

\item{states}{Vector of state space E (of length s).}
}
\value{
A numeric value giving the value of the AIC.
}
\description{
Generic function computing the Akaike Information Criterion of
the model \code{x}, with the list of sequences \code{sequences}.
}
