% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmnonparametric.R
\name{aic.smmnonparametric}
\alias{aic.smmnonparametric}
\title{Akaike Information Criterion (AIC)}
\usage{
\method{aic}{smmnonparametric}(x, sequences)
}
\arguments{
\item{x}{An object of class \link{smmnonparametric}.}

\item{sequences}{A list of vectors representing the sequences for which the
AIC criterion must be computed.}
}
\value{
Value of the AIC.
}
\description{
Computation of the Akaike Information Criterion.
}
