% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterPathCalc.r
\name{betterPathCalc}
\alias{betterPathCalc}
\title{Sparse Group Lasso Optimization}
\usage{
betterPathCalc(
  data,
  index,
  alpha = 0.95,
  min.frac = 0.05,
  nlam = 20,
  type = "linear"
)
}
\description{
Sparse Group Lasso Optimization
}
\keyword{internal}
