% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{bic}
\alias{bic}
\title{Bayesian Information Criterion (BIC)}
\usage{
bic(x, seq, E)
}
\arguments{
\item{x}{An object of class \link{markovmodel} or inheriting from the class smm.
(\link{smmnonparametric} or \link{smmparametric}).}

\item{seq}{A list of vectors representing the sequences for which the
BIC criterion must be computed.}

\item{E}{Vector of state space (of length S).}
}
\value{
A vector giving the value of the BIC for each sequence.
}
\description{
Computation of the Bayesian Information Criterion.
}
