% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computebic.R
\name{computebic}
\alias{computebic}
\title{Compute BIC scores for model selection}
\usage{
computebic(
  X,
  Y,
  M,
  a,
  b,
  c,
  tol = 10^(-10),
  max.iter = 100,
  grpgroup = c(1, rep(1:V + 1, 2)),
  threshold = 0
)
}
\arguments{
\item{X}{One-dimensional predictor}

\item{Y}{Outcome}

\item{M}{Multivariate mediator}

\item{tol}{(default -10^(-10)) convergence criterion}

\item{max.iter}{(default=100) maximum iteration}

\item{grpgroup}{(default=c(1,rep( 1:V +1,2)))}

\item{lambda}{(default=log(1+(1:50)/125)) tuning parameter for L1 penalization}

\item{penalty.factor}{(default=c(0,rep(sqrt(2),V))) give different weight of penalization for the 2V mediation paths.}
}
\value{
c directeffect

hatb Path b (M->Y given X) estimates

hata Path a (X->M) estimates

medest Mediation estimates (a*b)

alpha

lambda

nump Number of selected mediation paths
}
\description{
Compute BIC scores for model selection
}
\examples{
N=100
V=50
set.seed(1234)
a = rbinom(V,1,0.1)*5;b<-a
X = rnorm(N)
M =  X \%*\% t(a)+ matrix(rnorm(N*V),N,V)
Y =  X + M \%*\% b + rnorm(N)
}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{glmnet}
\keyword{highdimensional}
\keyword{mediation}
