% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.sparse.mediation.grplasso.R
\name{cv.sparse.mediation.grplasso}
\alias{cv.sparse.mediation.grplasso}
\title{Conduct K-fold cross validation for sparse mediation with group lasso with multiple tuning parameters}
\usage{
cv.sparse.mediation.grplasso(X, M, Y, tol = 10^(-10), K = 5,
  max.iter = 100, lambda = log(1 + (1:15)/40), grpgroup = c(1, rep(1:V +
  1, 2)), penalty.factor = c(0, rep(1, V)), threshold = 0,
  verbose = FALSE, multicore = 1, seednum = 1e+06)
}
\arguments{
\item{X}{One-dimensional predictor}

\item{M}{Multivariate mediator}

\item{Y}{Outcome}

\item{tol}{(default -10^(-10)) convergence criterion}

\item{K}{(default=5) number of cross-validation folds}

\item{max.iter}{(default=100) maximum iteration}

\item{lambda}{(default=log(1+(1:30)/100)) tuning parameter for L1 penalization}

\item{multicore}{(default=1) number of multicore}

\item{seednum}{(default=10000) seed number for cross validation}

\item{figure}{(defult=NULL) print figures for mean predictive errors by tuning parameters alpha and lambda}

\item{glmnet.penalty.factor}{(default=c(0,rep(1,2*V))) give different weight of penalization for the 2V mediation paths.}

\item{disply}{devault=FALSE}
}
\value{
cv.lambda: optimal lambda

cv.tau: optimal tau

cv.alpha: optimal tau

cv.mse: minimum MSE value

mse: Array of MSE, length(alpha) x length(lambda) x length (tau)

lambda: vector of lambda

tau: vector of tau used

alpha: vector of alpha used

z: cross-valication results
}
\description{
Conduct K-fold cross validation for sparse mediation with group lasso with multiple tuning parameters
}
\examples{
N=200
V=50
set.seed(1234)
a = rbinom(V,1,0.1)*5;b<-a
X = rnorm(N)
M =  X \%*\% t(a)+ matrix(rnorm(N*V),N,V)
Y =  as.vector(X + M \%*\% b + rnorm(N))
system.time(cvfit<-cv.sparse.mediation.grplasso(X, M, Y, K = 8,multicore = 4, seednum = 1e+06))
cvfit$cv.lambda
fit<-sparse.mediation.grplasso(X,M,Y,lambda = cvfit$cv.lambda)
nonzerogroups = 1-as.numeric((fit$hata!=0)+(fit$hatb!=0) ==0)
refit<-sparse.mediation.grplasso(X,M[,nonzerogroups==1],Y,lambda = 0)
}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{gglasso}
\keyword{hdlfpca}
