% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.sparse.mediation.msecomputing.R
\name{cv.sparse.mediation.msecomputing}
\alias{cv.sparse.mediation.msecomputing}
\title{Compare x to 1}
\usage{
cv.sparse.mediation.msecomputing(obj, Y.test, X.test, M.test)
}
\description{
Compare x to 1
}
\keyword{internal}
