% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.sparse.modmediation.grplasso.R
\name{cv.sparse.modmediation.grplasso}
\alias{cv.sparse.modmediation.grplasso}
\title{Conduct K-fold cross validation for sparse moderated mediation with group lasso}
\usage{
cv.sparse.modmediation.grplasso(
  X,
  M,
  Y,
  Z,
  tol = 10^(-5),
  K = 5,
  max.iter = 100,
  lambda = log(1 + (1:15)/40),
  grpgroup = c(rep(1, 3), rep(1:V + 1, 5)),
  penalty.factor = c(0, rep(1, V)),
  verbose = FALSE,
  multicore = 1,
  seednum = 1e+06
)
}
\arguments{
\item{X}{One-dimensional predictor}

\item{M}{Multivariate mediator}

\item{Y}{Outcome}

\item{tol}{(default -10^(-10)) convergence criterion}

\item{K}{(default=5) number of cross-validation folds}

\item{max.iter}{(default=100) maximum iteration}

\item{lambda}{(default=log(1+(1:30)/100)) tuning parameter for L1 penalization}

\item{grpgroup}{(default=c(rep(1,3),rep( 1:V +1,5)))}

\item{penalty.factor}{(default=c(0,rep(sqrt(2),V))) give different weight of penalization for the 2V mediation paths.}

\item{multicore}{(default=1) number of multicore}

\item{seednum}{(default=10000) seed number for cross validation}
}
\value{
cv.lambda: optimal lambda

cv.mse: minimum MSE value

mse: Array of MSE, length(alpha) x length(lambda) x length (tau)

lambda: vector of lambda

tau: vector of tau used

alpha: vector of alpha used

z: cross-valication results
}
\description{
Conduct K-fold cross validation for sparse moderated mediation with group lasso
}
\examples{
N=200
V=50
set.seed(1234)
a = rbinom(V,1,0.1)*5;b<-a
X = rnorm(N)
M =  X \%*\% t(a)+ matrix(rnorm(N*V),N,V)
Y =  as.vector(X + M \%*\% b + rnorm(N))
}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{gglasso}
\keyword{hdlfpca}
