% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{failureRateBMP}
\alias{failureRateBMP}
\title{BMP-Failure Rate Function}
\usage{
failureRateBMP(
  x,
  k,
  upstates = x$states,
  var = FALSE,
  epsilon = 0.001,
  klim = 10000
)
}
\arguments{
\item{x}{An object inheriting from the S3 class \code{smm} (an object of class
\link{smmparametric} or \link{smmnonparametric}).}

\item{k}{A positive integer giving the period \eqn{[0, k]} on which the
BMP-failure rate should be computed.}

\item{upstates}{Vector giving the subset of operational states \eqn{U}.}

\item{var}{Logical. If \code{TRUE} the asymptotic variance is computed.}

\item{epsilon}{Value of the reliability above which the latter is supposed
to be 0 because of computation errors (see Details).}

\item{klim}{Optional. The time horizon used to approximate the series in the
computation of the mean sojourn times vector \eqn{m} (cf.
\link{meanSojournTimes} function) for the asymptotic variance.}
}
\value{
A vector of length \eqn{k + 1} giving the values of the BMP-failure
rate for the period \eqn{[0,\dots,k]} if \code{var = FALSE}. If \code{var = TRUE},
a list containing the following components:
\itemize{
\item{x: }{a vector of length \eqn{k + 1} giving the values of the
BMP-failure rate for the period \eqn{[0,\dots,k]};}
\item{sigma2: }{a vector giving the asymptotic variance of the estimator
\eqn{\sigma_{\lambda}^{2}(k)}.}
}
}
\description{
Consider a system \eqn{S_{ystem}} starting to work at time
\eqn{k = 0}. The BMP-failure rate at time \eqn{k \in N} is the conditional
probability that the failure of the system occurs at time \eqn{k}, given
that the system has worked until time \eqn{k - 1}.
}
\details{
Consider a system (or a component) \eqn{S_{ystem}} whose possible
states during its evolution in time are \eqn{E = \{1,\dots,s\}}.
Denote by \eqn{U = \{1,\dots,s_1\}} the subset of operational states of
the system (the up states) and by \eqn{D = \{s_1 + 1,\dots, s\}} the
subset of failure states (the down states), with \eqn{0 < s_1 < s}
(obviously, \eqn{E = U \cup D} and \eqn{U \cap D = \emptyset},
\eqn{U \neq \emptyset,\ D \neq \emptyset}). One can think of the states
of \eqn{U} as different operating modes or performance levels of the
system, whereas the states of \eqn{D} can be seen as failures of the
systems with different modes.

We are interested in investigating the failure rate theory of a
discrete-time semi-Markov system \eqn{S_{ystem}}. Consequently, we suppose
that the evolution in time of the system is governed by an E-state space
semi-Markov chain \eqn{(Z_k)_{k \in N}}. The system starts to work at
instant \eqn{0} and the state of the system is given at each instant
\eqn{k \in N} by \eqn{Z_k}: the event \eqn{\{Z_k = i\}}, for a certain
\eqn{i \in U}, means that the system \eqn{S_{ystem}} is in operating mode
\eqn{i} at time \eqn{k}, whereas \eqn{\{Z_k = j\}}, for a certain
\eqn{j \in D}, means that the system is not operational at time \eqn{k}
due to the mode of failure \eqn{j} or that the system is under the
repairing mode \eqn{j}.

The BMP-failure rate at time \eqn{k \in N} is the conditional probability
that the failure of the system occurs at time \eqn{k}, given that the
system has worked until time \eqn{k - 1}.

Let \eqn{T_D} denote the first passage time in subset \eqn{D}, called
the lifetime of the system, i.e.,

\deqn{T_D := \textrm{inf}\{ n \in N;\ Z_n \in D\}\ \textrm{and}\ \textrm{inf}\ \emptyset := \infty.}

For a discrete-time semi-Markov system, the failure rate at time
\eqn{k \geq 1} has the expression:

\deqn{\lambda(k) := P(T_{D} = k | T_{D} \geq k)}

We can rewrite it as follows :

\deqn{\lambda(k) = 1 - \frac{R(k)}{R(k - 1)},\ \textrm{if } R(k - 1) \neq 0;\ \lambda(k) = 0, \textrm{otherwise}}

The failure rate at time \eqn{k = 0} is defined by \eqn{\lambda(0) := 1 - R(0)},
with \eqn{R} being the reliability function (see \link{reliability} function).

The calculation of the reliability \eqn{R} involves the computation of
many convolutions. It implies that the computation error, may be higher
(in value) than the "true" reliability itself for reliability close to 0.
In order to avoid inconsistent values of the BMP-failure rate, we use the
following formula:

\deqn{\lambda(k) = 1 - \frac{R(k)}{R(k - 1)},\ \textrm{if } R(k - 1) \geq \epsilon;\ \lambda(k) = 0, \textrm{otherwise}}

with \eqn{\epsilon}, the threshold, the parameter \code{epsilon} in the
function \code{failureRateBMP}.
}
