% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbic_int.R
\name{hbic_int}
\alias{hbic_int}
\title{Function to do HBIC computation for exposure by mediation interaction model}
\usage{
hbic_int(fit, fit.n)
}
\arguments{
\item{fit.n}{}
}
\value{

}
\description{
Function to do HBIC computation for exposure by mediation interaction model
}
