% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbic_med.R
\name{hbic_med}
\alias{hbic_med}
\title{Function to do HBIC computation for multiple mediations model}
\usage{
hbic_med(fit, fit.n)
}
\arguments{
\item{fit.n}{}
}
\value{

}
\description{
Function to do HBIC computation for multiple mediations model
}
