% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovmodel.R
\name{is.markovmodel}
\alias{is.markovmodel}
\title{Function to check if an object is of class \code{markovmodel}}
\usage{
is.markovmodel(x)
}
\arguments{
\item{x}{An arbitrary R object.}
}
\description{
\code{is.markovmodel} returns \code{TRUE} if \code{x} is an object of
class \code{markovmodel}.
}
