% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovmodelfit.R
\name{is.markovmodelfit}
\alias{is.markovmodelfit}
\title{Function to check if an object is of class \code{markovmodelfit}}
\usage{
is.markovmodelfit(x)
}
\arguments{
\item{x}{An arbitrary R object.}
}
\description{
\code{is.markovmodelfit} returns \code{TRUE} if \code{x} is an object of
class \code{markovmodelfit}.
}
