% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{is.mm}
\alias{is.mm}
\title{Function to check if an object is of class \code{mm}}
\usage{
is.mm(x)
}
\arguments{
\item{x}{An arbitrary R object.}
}
\description{
\code{is.mm} returns \code{TRUE} if \code{x} is an object of
class \code{mm}.
}
