% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovmodel.R
\name{loglik.markovmodel}
\alias{loglik.markovmodel}
\title{Loglikelihood}
\usage{
\method{loglik}{markovmodel}(x, sequences)
}
\arguments{
\item{x}{An object of class \link{markovmodel}.}

\item{sequences}{A list of vectors representing the sequences for which the
log-likelihood must be computed.}
}
\value{
A vector giving the value of the loglikelihood for each sequence.
}
\description{
Computation of the loglikelihood for a semi-Markov model
}
