% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{maintainability}
\alias{maintainability}
\title{Maintainability Function}
\usage{
maintainability(x, k, downstates = x$states)
}
\arguments{
\item{x}{An object inheriting from the S3 class \code{smm} (an object of class
\link{smmparametric} or \link{smmnonparametric}).}

\item{k}{A positive integer giving the period \eqn{[0, k]} on which the
maintainability should be computed.}

\item{downstates}{Vector giving the subset of non-operational states \eqn{D}.}
}
\value{
A vector of length \eqn{k + 1} giving the values of the maintainability
for the period \eqn{[0,\dots,k]}.
}
\description{
For a reparable system \eqn{S_{ystem}} for which the failure
occurs at time \eqn{k = 0}, its maintainability at time \eqn{k \in N} is
the probability that the system is repaired up to time \eqn{k}, given that
it has failed at time \eqn{k = 0}.
}
\details{
Consider a system (or a component) \eqn{S_{ystem}} whose possible
states during its evolution in time are \eqn{E = \{1,\dots,s\}}.
Denote by \eqn{U = \{1,\dots,s_1\}} the subset of operational states of
the system (the up states) and by \eqn{D = \{s_1 + 1,\dots, s\}} the
subset of failure states (the down states), with \eqn{0 < s_1 < s}
(obviously, \eqn{E = U \cup D} and \eqn{U \cap D = \emptyset},
\eqn{U \neq \emptyset,\ D \neq \emptyset}). One can think of the states
of \eqn{U} as different operating modes or performance levels of the
system, whereas the states of \eqn{D} can be seen as failures of the
systems with different modes.

We are interested in investigating the maintainability theory of a
discrete-time semi-Markov system \eqn{S_{ystem}}. Consequently, we suppose
that the evolution in time of the system is governed by an E-state space
semi-Markov chain \eqn{(Z_k)_{k \in N}}. The system starts to fail at
instant \eqn{0} and the state of the system is given at each instant
\eqn{k \in N} by \eqn{Z_k}: the event \eqn{\{Z_k = i\}}, for a certain
\eqn{i \in U}, means that the system \eqn{S_{ystem}} is in operating mode
\eqn{i} at time \eqn{k}, whereas \eqn{\{Z_k = j\}}, for a certain
\eqn{j \in D}, means that the system is not operational at time \eqn{k}
due to the mode of failure \eqn{j} or that the system is under the
repairing mode \eqn{j}.

Thus, we take \eqn{(\alpha_i)_{i \in U} = 0} and we denote by \eqn{T_U}
the first hitting time of subset \eqn{U}, called the duration of repair or
repair time, that is,

\deqn{T_U := \textrm{inf}\{ n \in N;\ Z_n \in U\}\ \textrm{and}\ \textrm{inf}\ \emptyset := \infty.}

The maintainability at time \eqn{k \in N} of a discrete-time semi-Markov
system is

\deqn{M(k) = P(T_U \leq k) = 1 - P(T_{U} \geq k) = 1 - P(Z_{n} \in D,\ n = 0,\dots,k).}
}
