% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovmodel.R
\name{markovmodel}
\alias{markovmodel}
\title{Markov model specification}
\usage{
markovmodel(states, init, ptrans, k = 1)
}
\arguments{
\item{states}{Vector of state space of length s.}

\item{init}{Vector of initial distribution of length s ^ k.}

\item{ptrans}{Matrix of transition probabilities of the embedded Markov chain
\eqn{J=(J_m)_{m}} of size sxs.}

\item{k}{Order of the Markov chain.}
}
\value{
An object of class \link{markovmodel}.
}
\description{
Creates a model specification of a Markov model.
}
\seealso{
\link{simulate}, \link{fitmarkovmodel}, \link{smmnonparametric}, \link{smmparametric}, \link{fitsemimarkovmodel}
}
