% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{meanSojournTimes}
\alias{meanSojournTimes}
\title{Mean Sojourn Times Function}
\usage{
meanSojournTimes(x, states = x$states, k = 10000)
}
\arguments{
\item{x}{An object inheriting from the S3 class \code{smm} (an object of class
\link{smmparametric} or \link{smmnonparametric}).}

\item{states}{Vector giving the states for which the mean sojourn time
should be computed. \code{states} is a subset of \eqn{E}.}

\item{k}{Optional. The time horizon used to approximate the series in the
computation of the mean sojourn times vector.}
}
\value{
A vector of length \eqn{\textrm{card}(E)} giving the values of the
mean sojourn times for each state \eqn{i \in E}.
}
\description{
Consider a system \eqn{S_{ystem}} starting to work at time
\eqn{k = 0}. The mean time to failure (MTTF) is defined as the mean
lifetime.
}
\details{
Consider a system (or a component) \eqn{S_{ystem}} whose possible
states during its evolution in time are \eqn{E = \{1,\dots,s\}}.

We are interested in investigating the mean sojourn times of a
discrete-time semi-Markov system \eqn{S_{ystem}}. Consequently, we suppose
that the evolution in time of the system is governed by an E-state space
semi-Markov chain \eqn{(Z_k)_{k \in N}}. The state of the system is given
at each instant \eqn{k \in N} by \eqn{Z_k}: the event \eqn{\{Z_k = i\}}.

Let \eqn{S = (S_{n})_{n \in N}} denote the successive time points when
state changes in \eqn{(Z_{n})_{n \in N}} occur.

The mean sojourn times vector is defined as follows:

\deqn{m_{i} = E[S_{1} | Z_{0} = j] = \sum_{n \geq 0} (1 - H_{j}(n)),\ i \in E}
}
