% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{mttr}
\alias{mttr}
\title{Mean Time To Repair (MTTR) Function}
\usage{
mttr(x, downstates = x$states, klim = 10000, var = FALSE)
}
\arguments{
\item{x}{An object inheriting from the S3 class \code{smm} (an object of class
\link{smmparametric} or \link{smmnonparametric}).}

\item{downstates}{Vector giving the subset of non-operational states \eqn{D}.}

\item{klim}{Optional. The time horizon used to approximate the series in the
computation of the mean sojourn times vector \eqn{m} (cf.
\link{meanSojournTimes} function) for the asymptotic
variance.}

\item{var}{Optional. A logical value indicating whether or not the variance
of the estimator should be computed.}
}
\value{
If \code{var = FALSE}, a vector of length \eqn{\textrm{card}(D)} giving
the values of the mean time to repair for each state \eqn{i \in D}. If
\code{var = TRUE}, a list containing the following components:
\itemize{
\item{x: }{a vector of length \eqn{\textrm{card}(D)} giving the values
of the mean time to repair for each state \eqn{i \in D}.}
\item{sigma2: }{the variances of the estimator for each estimation of the
mean time to repair \eqn{\sigma_{MTTR_{i}}^{2}(k)};}
}
}
\description{
Consider a system \eqn{S_{ystem}} starting to fail at time
\eqn{k = 0}. The mean time to repair (MTTR) is defined as the mean of the
repair duration.
}
\details{
Consider a system (or a component) \eqn{S_{ystem}} whose possible
states during its evolution in time are \eqn{E = \{1,\dots,s\}}.
Denote by \eqn{U = \{1,\dots,s_1\}} the subset of operational states of
the system (the up states) and by \eqn{D = \{s_1 + 1,\dots, s\}} the
subset of failure states (the down states), with \eqn{0 < s_1 < s}
(obviously, \eqn{E = U \cup D} and \eqn{U \cap D = \emptyset},
\eqn{U \neq \emptyset,\ D \neq \emptyset}). One can think of the states
of \eqn{U} as different operating modes or performance levels of the
system, whereas the states of \eqn{D} can be seen as failures of the
systems with different modes.

We are interested in investigating the mean time to repair theory of a
discrete-time semi-Markov system \eqn{S_{ystem}}. Consequently, we suppose
that the evolution in time of the system is governed by an E-state space
semi-Markov chain \eqn{(Z_k)_{k \in N}}. The system starts to fail at
instant \eqn{0} and the state of the system is given at each instant
\eqn{k \in N} by \eqn{Z_k}: the event \eqn{\{Z_k = i\}}, for a certain
\eqn{i \in U}, means that the system \eqn{S_{ystem}} is in operating mode
\eqn{i} at time \eqn{k}, whereas \eqn{\{Z_k = j\}}, for a certain
\eqn{j \in D}, means that the system is not operational at time \eqn{k}
due to the mode of failure \eqn{j} or that the system is under the
repairing mode \eqn{j}.

Let \eqn{T_U} denote the first passage time in subset \eqn{U}, called the
duration of repair or repair time, i.e.,

\deqn{T_U := \textrm{inf}\{ n \in N;\ Z_n \in U\}\ \textrm{and}\ \textrm{inf}\ \emptyset := \infty.}

The mean time to repair (MTTR) is defined as the mean of the repair
duration, i.e., the expectation of the hitting time to up set \eqn{U},

\deqn{MTTR = E[T_{U}]}
}
