% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneDim.R
\name{oneDim}
\alias{oneDim}
\title{Sparse Group Lasso Optimization}
\usage{
oneDim(
  data,
  index,
  thresh = 1e-04,
  nlam = 20,
  lambdas = NULL,
  beta.naught = rep(0, ncol(data$x)),
  inner.iter = 100,
  outer.iter = 100,
  outer.thresh = 1e-04,
  gamma = 0.8,
  step = 1,
  reset = 10,
  alpha = 0.95,
  min.frac = 0.05,
  verbose = FALSE
)
}
\description{
Sparse Group Lasso Optimization
}
\keyword{internal}
