% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmnonparametric.R
\name{plot.smmnonparametric}
\alias{plot.smmnonparametric}
\title{Plot function for an object of class smmnonparametric}
\usage{
\method{plot}{smmnonparametric}(x, i = 1, j = 1, klim = NULL, ...)
}
\arguments{
\item{x}{An object of class \link{smmnonparametric}.}

\item{i}{An integer giving the current state in the following cases:
\code{type.sojourn = "fij"} or \code{type.sojourn = "fi"}, otherwise, \code{i} is
ignored.}

\item{j}{An integer giving the next state in the following cases:
\code{type.sojourn = "fij"} or \code{type.sojourn = "fj"}, otherwise, \code{j} is
ignored.}

\item{klim}{An integer giving the limit value for which the density will be
plotted. If \code{klim} is \code{NULL}, then quantile or order 0.95 is used.}

\item{...}{Arguments passed to plot.}
}
\description{
Displays the densities for the conditional sojourn time
distributions depending on the current state \code{i} and on the next state
\code{j}.
}
