% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.markovmodel.R
\name{simulate.markovmodel}
\alias{simulate.markovmodel}
\title{Simulates k-th order Markov chains}
\usage{
\method{simulate}{markovmodel}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{An object of class \link{markovmodel}.}

\item{nsim}{An integer or vector of integers (for multiple sequences)
specifying the length of the sequence(s).}

\item{seed}{\code{seed} for the random number generator.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of vectors representing the sequences.
}
\description{
Simulates k-th order Markov chains.
}
\details{
If \code{nsim} is a single integer then a chain of that length is
produced. If \code{nsim} is a vector of integers, then \code{length(nsim)}
sequences are generated with respective lengths.
}
\seealso{
\link{fitmarkovmodel}, \link{smmnonparametric}, \link{smmparametric}, \link{fitsemimarkovmodel}
}
