% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{simulate.smm}
\alias{simulate.smm}
\title{Simulates semi-Markov chains}
\usage{
\method{simulate}{smm}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{An object inheriting from the S3 class \code{smm} (an object of
class \link{smmparametric} or \link{smmnonparametric}).}

\item{nsim}{An integer or vector of integers (for multiple sequences)
specifying the length of the sequence(s).}

\item{seed}{\code{seed} for the random number generator.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of vectors representing the sequences.
}
\description{
Simulates sequences from a semi-Markov model.
}
\details{
If \code{nsim} is a single integer then a chain of that length is
produced. If \code{nsim} is a vector of integers, then \code{length(nsim)}
sequences are generated with respective lengths.
}
\seealso{
\link{smmparametric}, \link{smmnonparametric}, \link{fitsemimarkovmodel}
}
