% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.mediation.grplasso.largep_omega.R
\name{sparse.mediation.grplasso.largep_omega}
\alias{sparse.mediation.grplasso.largep_omega}
\title{Conduct sparse mediation for large p ( p > n) with grouplasso penalty and fast computation of inverse matrix}
\usage{
sparse.mediation.grplasso.largep_omega(
  X,
  M,
  Y,
  tol = 10^(-10),
  max.iter = 100,
  lambda1 = exp(-8:0),
  lambda2 = exp(seq(0, 0.8 * log(ncol(M)), length = 4)),
  grpgroup = c(1, rep(1:(ncol(M)) + 1, 2)),
  penalty.factor = c(0, rep(1, ncol(M))),
  verbose = FALSE,
  Omega.out = FALSE
)
}
\arguments{
\item{X}{One-dimensional predictor}

\item{M}{Multivariate mediator}

\item{Y}{Outcome}

\item{tol}{(default -10^(-10)) convergence criterion}

\item{max.iter}{(default=100) maximum iteration}

\item{grpgroup}{(default=c(1,rep( 1:V +1,2)))}

\item{penalty.factor}{(default=c(0,rep(sqrt(2),V))) give different weight of penalization for the 2V mediation paths.}

\item{lambda}{(default=log(1+(1:50)/125)) tuning parameter for L1 penalization}
}
\value{
c directeffect

hatb Path b (M->Y given X) estimates

hata Path a (X->M) estimates

medest Mediation estimates (a*b)

alpha

lambda

nump Number of selected mediation paths
}
\description{
Fit a mediation model via penalized maximum likelihood and structural equation model.
The regularization path is computed for the grouplasso penalty at a grid of
values for the regularization parameter lambda. Currently, mediation analysis is developed based on gaussian assumption.
}
\details{
Multiple Mediaton Model:
(1) M = Xa + e1
(2) Y = Xc' + Mb + e2
And in the optimization, we do not regularize c', due to the assumption of partial mediation.
}
\examples{
N=100
V=300
set.seed(1234)
a = c(rep(1,3),rep(0,V-3))*5;b<-a
X = rnorm(N)
M =  X \%*\% t(a)+ matrix(rnorm(N*V),N,V)
Y =  X + M \%*\% b + rnorm(N)
fit=sparse.mediation.grplasso.largep_omega(X,M,Y,verbose=FALSE)
}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{L1penalization}
\keyword{highdimensional}
\keyword{mediation}
