% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.mediation.grplasso.R,
%   R/sparse.modmediation.grplasso.R
\name{sparse.mediation.grplasso}
\alias{sparse.mediation.grplasso}
\alias{sparse.mediation.grplasso}
\title{Conduct sparse mediation with group LASSO (Old version)}
\usage{
sparse.mediation.grplasso(X, M, Y, tol = 10^(-10), max.iter = 100,
  lambda = log(1 + (1:50)/125), grpgroup = c(rep(1, 3), rep(1:V + 1, 5)),
  penalty.factor = c(0, rep(1, V)), threshold = 0, verbose = FALSE)

sparse.mediation.grplasso(X, M, Y, tol = 10^(-10), max.iter = 100,
  lambda = log(1 + (1:50)/125), grpgroup = c(rep(1, 3), rep(1:V + 1, 5)),
  penalty.factor = c(0, rep(1, V)), threshold = 0, verbose = FALSE)
}
\arguments{
\item{X}{One-dimensional predictor}

\item{M}{Multivariate mediator}

\item{Y}{Outcome}

\item{tol}{(default -10^(-10)) convergence criterion}

\item{max.iter}{(default=100) maximum iteration}

\item{lambda}{(default=log(1+(1:50)/125)) tuning parameter for L1 penalization}

\item{grpgroup}{(default=c(1,rep( 1:V +1,2)))}

\item{penalty.factor}{(default=c(0,rep(sqrt(2),V))) give different weight of penalization for the 2V mediation paths.}

\item{X}{One-dimensional predictor}

\item{M}{Multivariate mediator}

\item{Y}{Outcome}

\item{Z}{Scalar moderator}

\item{tol}{(default -10^(-10)) convergence criterion}

\item{max.iter}{(default=100) maximum iteration}

\item{lambda}{(default=log(1+(1:50)/125)) tuning parameter for L1 penalization}

\item{grpgroup}{(default=c(rep(1,3),rep( 1:V +1,5)))}

\item{penalty.factor}{(default=c(0,rep(sqrt(2),V))) give different weight of penalization for the 2V mediation paths.}
}
\value{
c directeffect

hatb Path b (M->Y given X) estimates

hata Path a (X->M) estimates

medest Mediation estimates (a*b)

alpha

lambda

nump Number of selected mediation paths

c directeffect

hatb Path b (M->Y given X) estimates

hata Path a (X->M) estimates

medest Mediation estimates (a*b)

alpha

lambda

nump Number of selected mediation paths
}
\description{
Fit a mediation model via penalized maximum likelihood and structural equation model.
The regularization path is computed for the lasso or elasticnet penalty at a grid of
values for the regularization parameter lambda. Currently, mediation analysis is developed based on gaussian assumption.

Fit a mediation model via penalized maximum likelihood and structural equation model.
The regularization path is computed for the lasso or elasticnet penalty at a grid of
values for the regularization parameter lambda. Currently, mediation analysis is developed based on gaussian assumption.
}
\details{
Multiple Mediaton Model:
(1) M = Xa + e1
(2) Y = Xc' + Mb + e2
And in the optimization, we do not regularize c', due to the assumption of partial mediation.

Multiple Mediaton Model:
(1) M = Xa + e1
(2) Y = Xc' + Mb + e2
And in the optimization, we do not regularize c', due to the assumption of partial mediation.
}
\examples{
N=100
V=50
set.seed(1234)
a = rbinom(V,1,0.1)*5;b<-a
X = rnorm(N)
M =  X \%*\% t(a)+ matrix(rnorm(N*V),N,V)
Y =  X + M \%*\% b + rnorm(N)
fit=sparse.mediation.grplasso(X,M,Y,verbose=FALSE, lambda = log(1+(1:20)/50))
N=100
V=50
set.seed(1234)
a = rbinom(V,1,0.1)*5;b<-a
X = rnorm(N)
M =  X \%*\% t(a)+ matrix(rnorm(N*V),N,V)
Y =  X + M \%*\% b + rnorm(N)
fit=sparse.mediation.grplasso(X,M,Y,verbose=FALSE, lambda = log(1+(1:20)/50))
}
\references{
TBA

TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}

Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{glmnet}
\keyword{highdimensional}
\keyword{mediation}
