% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.modmediation.grplasso.R
\name{sparse.modmediation.grplasso}
\alias{sparse.modmediation.grplasso}
\title{Conduct sparse moderated mediation with group LASSO}
\usage{
sparse.modmediation.grplasso(
  X,
  M,
  Y,
  Z,
  tol = 10^(-10),
  max.iter = 100,
  lambda = log(1 + (1:10)/10),
  X.cont = TRUE,
  grpgroup = c(1, rep(1:(ncol(M)) + 1, 2)),
  penalty.factor = c(0, rep(1, ncol(M))),
  threshold = 0,
  verbose = FALSE
)
}
\arguments{
\item{X}{One-dimensional predictor}

\item{M}{Multivariate mediator}

\item{Y}{Outcome}

\item{Z}{Scalar moderator}

\item{tol}{(default -10^(-10)) convergence criterion}

\item{max.iter}{(default=100) maximum iteration}

\item{lambda}{(default=log(1+(1:50)/125)) tuning parameter for L1 penalization}

\item{X.cont}{(defult=TRUE) Related to standardization. If X is categorical data, standardiziagn won't apply.}

\item{grpgroup}{(default=c(rep(1,3),rep( 1:V +1,5)))}

\item{penalty.factor}{(default=c(0,rep(sqrt(2),V))) give different weight of penalization for the 2V mediation paths.}
}
\value{
c directeffect

hatb Path b (M->Y given X) estimates

hata Path a (X->M) estimates

medest Mediation estimates (a*b)

alpha

lambda

nump Number of selected mediation paths
}
\description{
Fit a mediation model via penalized maximum likelihood and structural equation model.
The regularization path is computed for the lasso or elasticnet penalty at a grid of
values for the regularization parameter lambda. Currently, mediation analysis is developed based on gaussian assumption.
}
\details{
Multiple Mediaton Model:
(1) M = Xa1 + XZ a2 + z a3 + e1
(2) Y = Xc1 + XZ c2  + Mb1 + MZ b2 + Z c3 + e2
And in the optimization, we do not regularize c', due to the assumption of partial mediation.
}
\examples{
N=100
V=50
set.seed(1234)
a = c(rep(1,3),  rep(0,V-3))*5;b<-a
X = rnorm(N)
M =  X \%*\% t(a)+ matrix(rnorm(N*V),N,V)
Y =  X + M \%*\% b + rnorm(N)
}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{glmnet}
\keyword{highdimensional}
\keyword{mediation}
