% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.txtmedint.sgrplasso.largep_omega.R
\name{sparse.txtmedint.sgrlasso.largep_omega}
\alias{sparse.txtmedint.sgrlasso.largep_omega}
\title{Conduct sparse mediation for large p ( p > n)}
\usage{
sparse.txtmedint.sgrlasso.largep_omega(
  X,
  M,
  Y,
  X.scale = TRUE,
  tol = 10^(-10),
  max.iter = 10,
  lambda1 = exp(seq(1, -3, length = 10)),
  lambda2 = exp(-1),
  alpha = c(0.95, 0.5),
  group.penalty.factor = c(0, rep(1, ncol(M))),
  penalty.factor = c(0, rep(1, ncol(M) * 3)),
  verbose = FALSE,
  Omega.out = FALSE,
  threshold = 0,
  non.zeros.stop = ncol(M)
)
}
\arguments{
\item{X}{One-dimensional predictor}

\item{M}{Multivariate mediator}

\item{Y}{Outcome}

\item{tol}{(default -10^(-10)) convergence criterion}

\item{max.iter}{(default=100) maximum iteration}

\item{lambda1}{(default=seq(0.02,0.4,length=5)) tuning parameter for regression coefficient L1 penalization}

\item{lambda2}{Tuning parameter for Covariance matrix L1 penalization}

\item{alpha}{(defult=1) tuning parameter for L2 penalization}

\item{group.penalty.factor}{(V+1)-dimensional group penalty factor vector. If a user does not want to penalize mediator, specify 0 otherwise 1. The first element is the direct effect followed by V-mediators. The default value is c(0,rep(1,V)).}

\item{penalty.factor}{(1+3*V)-dimensional sparsity penalty factor vector.}

\item{verbose}{(default=FALSE) print progress}

\item{Omega.out}{(defult=FALSE) output Omega estimates (beta version WIP.)}

\item{threshold}{(default=10^(-5))}

\item{non.zeros.stop}{(default=ncol(M)) When to stop the regularization path.}
}
\value{
c directeffect

hatb Path b (M->Y given X) estimates

hata Path a (X->M) estimates

medest Mediation estimates (a*b)

alpha

lambda1 Tuning parameters for regression coefficients

lambda2 Tuning parameters for inversed covariance matrix (Omega)

nump Number of selected mediation paths

Omega Estimated covariance matrix of the mediator
}
\description{
Sparse mediation with sparse group lasso (for mediation paths)
and sparse precision matrix estimation using fast computation of inverse matrix
}
\details{
Fit a mediation model via penalized maximum likelihood and structural equation model.
The regularization path is computed for the lasso or elasticnet penalty at a grid of
values for the regularization parameter lambda. Currently, mediation analysis is developed based on gaussian assumption.

Multiple Mediaton Model:
(1) M = Xa + e1
(2) Y = Xc' + Mb + e2
And in the optimization, we do not regularize c', due to the assumption of partial mediation.
}
\examples{
library(sparsemediation)
N=200
V=100
set.seed(1234)
a = c(rep(0,V-3),rep(1,3))*5;
b1=b2=rep(0,V)
b1[V]<-5
b2[V]<-5
X = rnorm(N)
M =  X \%*\% t(a)+ matrix(rnorm(N*V),N,V)
Y =  X + M \%*\% b1 + (X*M) \%*\% b2 +  rnorm(N)
#fit=sparse.txtmedint.sgrlasso.largep_omega(X,M,Y)

}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{L1penalization}
\keyword{highdimensional}
\keyword{mediation}
