% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovmodel.R
\name{aic.markovmodel}
\alias{aic.markovmodel}
\title{Akaike Information Criterion (AIC)}
\usage{
\method{aic}{markovmodel}(x, sequences)
}
\arguments{
\item{x}{An object of class \link{markovmodel}.}

\item{sequences}{A list of vectors representing the sequences for which the
AIC will be computed based on \code{x}.}
}
\value{
Value of the AIC.
}
\description{
Computation of the Akaike Information Criterion.
}
