% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{aic}
\alias{aic}
\title{Akaike Information Criterion (AIC)}
\usage{
aic(x, sequences)
}
\arguments{
\item{x}{An object inheriting from the S3 class \code{smm} (an object of class
\link{smmparametric} or \link{smmnonparametric}).}

\item{sequences}{A list of vectors representing the sequences for which the
AIC will be computed based on \code{x}.}
}
\value{
Value of the AIC.
}
\description{
Generic function computing the Akaike Information Criterion of
the model \code{x}, with the list of sequences \code{sequences}.
}
