% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic.R
\name{aic}
\alias{aic}
\title{Akaike Information Criterion (AIC)}
\usage{
aic(x, seq, E)
}
\arguments{
\item{x}{An object of class \link{markovmodel} or inheriting from the class smm.
(\link{smmnonparametric} or \link{smmparametric}).}

\item{seq}{A list of vectors representing the sequences for which the
AIC criterion must be computed.}

\item{E}{Vector of state space (of length S).}
}
\value{
A vector giving the value of the AIC for each sequence.
}
\description{
Computation of the Akaike Information Criterion.
}
