% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmparametric.R
\name{aic.smmparametric}
\alias{aic.smmparametric}
\title{Akaike Information Criterion (AIC)}
\usage{
\method{aic}{smmparametric}(x, sequences, states)
}
\arguments{
\item{x}{An object of class \link{smmparametric}.}

\item{sequences}{A list of vectors representing the sequences for which the
AIC criterion must be computed.}

\item{states}{Vector of state space (of length s).}
}
\value{
A numeric value giving the value of the AIC.
}
\description{
Computation of the Akaike Information Criterion.
}
