% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmparametric.R
\name{aic.smmparametric}
\alias{aic.smmparametric}
\title{Akaike Information Criterion (AIC)}
\usage{
\method{aic}{smmparametric}(x, sequences)
}
\arguments{
\item{x}{An object of class \link{smmparametric}.}

\item{sequences}{A list of vectors representing the sequences for which the
AIC will be computed based on \code{x}.}
}
\value{
Value of the AIC.
}
\description{
Computation of the Akaike Information Criterion.
}
