% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovmodel.R
\name{bic.markovmodel}
\alias{bic.markovmodel}
\title{Bayesian Information Criterion (BIC)}
\usage{
\method{bic}{markovmodel}(x, sequences)
}
\arguments{
\item{x}{An object of class \link{markovmodel}.}

\item{sequences}{A list of vectors representing the sequences for which the
BIC criterion must be computed.}
}
\value{
A numeric value giving the value of the BIC.
}
\description{
Computation of the Bayesian Information Criterion.
}
