% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovmodel.R
\name{bic.markovmodel}
\alias{bic.markovmodel}
\title{Bayesian Information Criterion (BIC)}
\usage{
\method{bic}{markovmodel}(x, seq, E)
}
\arguments{
\item{x}{An object of class \link{markovmodel}.}

\item{seq}{A list of vectors representing the sequences for which the
BIC criterion must be computed.}

\item{E}{Vector of state space (of length S).}
}
\value{
A vector giving the value of the BIC for each sequence.
}
\description{
Computation of the Bayesian Information Criterion.
}
\examples{
E <- c("a", "c", "g", "t")
S <- length(E)
vect.init <- c(1 / 4, 1 / 4, 1 / 4, 1 / 4)
k <- 2
p <- matrix(0.25, nrow = S ^ k, ncol = S)

# Specify the Markov model
markov1 <- markovmodel(E = E, init = vect.init, ptrans = p, k = k)
markov1

}
