% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{bic}
\alias{bic}
\title{Bayesian Information Criterion (BIC)}
\usage{
bic(x, seq, E)
}
\arguments{
\item{x}{An object for which the logikelihood can be computed.}

\item{seq}{A list of vectors representing the sequences for which the
BIC criterion must be computed.}

\item{E}{Vector of state space (of length S).}
}
\value{
A numeric value giving the value of the BIC.
}
\description{
Generic function computing the Bayesian Information Criterion
of the model \code{x}, with the list of sequences \code{seq}.
}
