% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{bic}
\alias{bic}
\title{Bayesian Information Criterion (BIC)}
\usage{
bic(x, sequences = NULL)
}
\arguments{
\item{x}{An object for which there exists a \code{loglik} attribute if
\code{sequences = NULL} or a \code{loglik} method otherwise.}

\item{sequences}{Optional. A list of vectors representing the sequences for
which the AIC will be computed based on \code{x} using the method \code{loglik}.}
}
\value{
Value of the BIC.
}
\description{
Generic function computing the Bayesian Information Criterion
of the model \code{x}, with the list of sequences \code{sequences}.
}
