% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmnonparametric.R
\name{bic.smmnonparametric}
\alias{bic.smmnonparametric}
\title{Bayesian Information Criterion (BIC)}
\usage{
\method{bic}{smmnonparametric}(x, seq, E)
}
\arguments{
\item{x}{An object of class \link{smmnonparametric}.}

\item{seq}{A list of vectors representing the sequences for which the
BIC criterion must be computed.}

\item{E}{Vector of state space (of length S).}
}
\value{
A numeric value giving the value of the BIC.
}
\description{
Computation of the Bayesian Information Criterion.
}
