% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmparametric.R
\name{bic.smmparametric}
\alias{bic.smmparametric}
\title{Bayesian Information Criterion (BIC)}
\usage{
\method{bic}{smmparametric}(x, sequences)
}
\arguments{
\item{x}{An object of class \link{smmparametric}.}

\item{sequences}{A list of vectors representing the sequences for which the
BIC criterion must be computed.}
}
\value{
Value of the BIC.
}
\description{
Computation of the Bayesian Information Criterion.
}
