% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmparametric.R
\name{bic.smmparametric}
\alias{bic.smmparametric}
\title{Bayesian Information Criterion (BIC)}
\usage{
\method{bic}{smmparametric}(x, seq, E)
}
\arguments{
\item{x}{An object of class \link{smmparametric}.}

\item{seq}{A list of vectors representing the sequences for which the
BIC criterion must be computed.}

\item{E}{Vector of state space (of length S).}
}
\value{
A numeric value giving the value of the BIC.
}
\description{
Computation of the Bayesian Information Criterion.
}
