% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{failureRateRG}
\alias{failureRateRG}
\title{RG-Failure Rate Function}
\usage{
failureRateRG(
  x,
  k,
  upstates = x$states,
  level = 0.95,
  epsilon = 0.001,
  klim = 10000
)
}
\arguments{
\item{x}{An object of S3 class \code{smmfit} or \code{smm}.}

\item{k}{A positive integer giving the period \eqn{[0, k]} on which the
RG-failure rate should be computed.}

\item{upstates}{Vector giving the subset of operational states \eqn{U}.}

\item{level}{Confidence level of the asymptotic confidence interval. Helpful
for an object \code{x} of class \code{smmfit}.}

\item{epsilon}{Value of the reliability above which the latter is supposed
to be 0 because of computation errors (see Details).}

\item{klim}{Optional. The time horizon used to approximate the series in the
computation of the mean sojourn times vector \eqn{m} (cf.
\link{meanSojournTimes} function) for the asymptotic variance.}
}
\value{
A matrix with \eqn{k + 1} rows, and with columns giving values of
the RG-failure rate, variances, lower and upper asymptotic confidence
limits (if \code{x} is an object of class \code{smmfit}).
}
\description{
Discrete-time adapted failure rate, proposed by D. Roy and
R. Gupta. Classification of discrete lives. Microelectronics Reliability,
32(10):1459--1473, 1992.
We call it the RG-failure rate and denote it by \eqn{r(k),\ k \in N}.
}
\details{
Expressing \eqn{r(k)} in terms of the reliability \eqn{R} we obtain
that the RG-failure rate function for a discrete-time system is given by:

\deqn{r(k) = - \ln \frac{R(k)}{R(k - 1)},\ \textrm{if } k \geq 1;\ r(k) = - \ln R(0),\ \textrm{if } k = 0}

for \eqn{R(k) \neq 0}. If \eqn{R(k) = 0}, we set \eqn{r(k) := 0}.

Note that the RG-failure rate is related to the BMP-failure rate
(see \link{failureRateBMP} function) by:

\deqn{r(k) = - \ln (1 - \lambda(k)),\ k \in N}

The computation of the RG-failure rate is based on the \link{failureRateBMP}
function (See \link{failureRateBMP} for details about the parameter \code{epsilon}).
}
