% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{failureRateRG}
\alias{failureRateRG}
\title{RG-Failure Rate Function}
\usage{
failureRateRG(x, k, alpha = x$init, upstates = x$states)
}
\arguments{
\item{x}{An object inheriting from the S3 class \code{smm} (an object of class
\link{smmparametric} or \link{smmnonparametric}).}

\item{k}{A positive integer giving the period \eqn{[0, k]} on which the
RG-Failure Rate should be computed.}

\item{alpha}{Vector of initial distribution \eqn{(\alpha_i)_{i \in U}} of
length \eqn{\textrm{card}(U) = s1}.}

\item{upstates}{Vector giving the subset of operational states \eqn{U}.}
}
\value{
A vector of length \eqn{k + 1} giving the values of the RG-Failure
Rate for the period \eqn{[0,\dots,k]}.
}
\description{
Discrete-time adapted failure rate, proposed by Roy and Gupta (1992).
We call it the RG-failure rate and denote it by \eqn{r(k),\ k \in N}.
}
