% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmarkovmodel.R
\name{fitmarkovmodel}
\alias{fitmarkovmodel}
\title{Estimation of a k-th order Markov chain}
\usage{
fitmarkovmodel(sequences, states, k = 1)
}
\arguments{
\item{sequences}{A list of vectors representing the sequences.}

\item{states}{Vector of state space (of length s).}

\item{k}{Order of the Markov chain.}
}
\value{
An object of class \link{markovmodel}.
}
\description{
Estimation of the transition matrix and initial law of a k-th
order Markov chain starting from one or several sequences.
}
\details{
Let \eqn{X_1, X_2, ..., X_n} be a trajectory of length \eqn{n} of
the Markov chain \eqn{X = (X_m)_{m \in N}} of order \eqn{k = 1} with
transition matrix \eqn{p_{trans}(i,j) = P(X_{m+1} = j | X_m = i)}. The
estimation of the transition matrix is \eqn{\widehat{p_{trans}}(i,j) = \frac{N_{ij}}{N_{i.}}},
where \eqn{N_{ij}} is the number of transitions from state \eqn{i} to state
\eqn{j} and \eqn{N_{i.}} is the number of transition from state \eqn{i}
to any state. For \eqn{k > 1} we have similar expressions.

The initial distribution of a k-th order Markov chain is defined as
\eqn{\mu_i = P(X_1 = i)}. An estimation of the initial law for a first order
Markov chain is assumed to be the estimation of the stationary distribution.
If the order of the Markov is greater than 1, then an estimation of the
initial law is \eqn{\widehat{\mu_i} = \frac{N_i}{N}}, where \eqn{N_i} is the number occurences
of state \eqn{i} in the sequences and \eqn{N} is the sum of the sequence
lengths.
}
\examples{
states <- c("a", "c", "g", "t")
s <- length(states)
vect.init <- c(1 / 4, 1 / 4, 1 / 4, 1 / 4)
k <- 2
p <- matrix(0.25, nrow = s ^ k, ncol = s)

# Specify the Markov model
markov1 <- markovmodel(states = states, init = vect.init, ptrans = p, k = k)

seq1 <- simulate(object = markov1, nsim = c(1000, 10000, 2000), seed = 150)

est <- fitmarkovmodel(sequences = seq1, states = states, k = 2)

}
\seealso{
\link{markovmodel}, \link{simulate}, \link{smmnonparametric}, \link{smmparametric}, \link{fitsemimarkovmodel}
}
