% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovmodel.R
\name{loglik.markovmodel}
\alias{loglik.markovmodel}
\title{Loglikelihood}
\usage{
\method{loglik}{markovmodel}(x, sequences)
}
\arguments{
\item{x}{An object of class \link{markovmodel}.}

\item{sequences}{A list of vectors representing the sequences for which the
log-likelihood will be computed based on \code{x}.}
}
\value{
Value of the log-likelihood.
}
\description{
Computation of the log-likelihood for a Markov model
}
