% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Loglikelihood}
\usage{
loglik(x, sequences, states)
}
\arguments{
\item{x}{An object for which the logikelihood can be computed.}

\item{sequences}{A list of vectors representing the sequences for which the
log-likelihood must be computed.}

\item{states}{Vector of state space (of length s).}
}
\value{
A vector giving the value of the loglikelihood for each sequence.
}
\description{
Generic function computing the loglikelihood of the model \code{x},
with the list of sequences \code{sequences}.
}
