% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Log-likelihood Function}
\usage{
loglik(x, sequences)
}
\arguments{
\item{x}{An object for which the log-likelihood can be computed.}

\item{sequences}{A list of vectors representing the sequences for which the
log-likelihood will be computed based on \code{x}.}
}
\value{
Value of the log-likelihood.
}
\description{
Generic function computing the log-likelihood of the model \code{x},
with the list of sequences \code{sequences}.
}
