% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Loglikelihood}
\usage{
loglik(x, seq, E)
}
\arguments{
\item{x}{An object of class \link{markovmodel} or inheriting from the class smm.
(\link{smmnonparametric} or \link{smmparametric}).}

\item{seq}{A list of vectors representing the sequences for which the
log-likelihood must be computed.}

\item{E}{Vector of state space (of length S).}
}
\value{
A vector giving the value of the loglikelihood for each sequence.
}
\description{
Computation of the loglikelihood for a Markov or a semi-Markov
model
}
