% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmnonparametric.R
\name{loglik.smmnonparametric}
\alias{loglik.smmnonparametric}
\title{Log-likelihood Function}
\usage{
\method{loglik}{smmnonparametric}(x, sequences)
}
\arguments{
\item{x}{An object of class \link{smmnonparametric}.}

\item{sequences}{A list of vectors representing the sequences for which the
log-likelihood will be computed based on \code{x}.}
}
\value{
Value of the log-likelihood.
}
\description{
Computation of the log-likelihood for a semi-Markov model
}
