% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmnonparametric.R
\name{loglik.smmnonparametric}
\alias{loglik.smmnonparametric}
\title{Loglikelihood}
\usage{
\method{loglik}{smmnonparametric}(x, sequences)
}
\arguments{
\item{x}{An object of class \link{smmnonparametric}.}

\item{sequences}{A list of vectors representing the sequences for which the
log-likelihood must be computed.}
}
\value{
Value of the loglikelihood.
}
\description{
Computation of the loglikelihood for a semi-Markov model
}
