% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmparametric.R
\name{loglik.smmparametric}
\alias{loglik.smmparametric}
\title{Loglikelihood}
\usage{
\method{loglik}{smmparametric}(x, sequences, states)
}
\arguments{
\item{x}{An object of class \link{smmparametric}.}

\item{sequences}{A list of vectors representing the sequences for which the
log-likelihood must be computed.}

\item{states}{Vector of state space (of length s).}
}
\value{
A vector giving the value of the loglikelihood for each sequence.
}
\description{
Computation of the loglikelihood for a semi-Markov model
}
