% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.smmparametric.R
\name{simulate.smmparametric}
\alias{simulate.smmparametric}
\title{Simulates semi-Markov chains}
\usage{
\method{simulate}{smmparametric}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{An object of class \link{smmparametric}.}

\item{nsim}{An integer or vector of integers (for multiple sequences)
specifying the length of the sequence(s).}

\item{seed}{\code{seed} for the random number generator.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of vectors representing the sequences.
}
\description{
Simulates sequences from a parametric semi-Markov model.
}
\details{
If \code{nsim} is a single integer then a chain of that length is
produced. If \code{nsim} is a vector of integers, then \code{length(nsim)}
sequences are generated with respective lengths.
}
\examples{
states <- c("a", "c", "g", "t")
s <- length(states)

# Creation of the initial distribution
vect.init <- c(1 / 4, 1 / 4, 1 / 4, 1 / 4)

# Creation of transition matrix
pij <- matrix(c(0, 0.2, 0.5, 0.3, 
                0.2, 0, 0.3, 0.5, 
                0.3, 0.5, 0, 0.2, 
                0.4, 0.2, 0.4, 0), 
              ncol = s, byrow = TRUE)

# Creation of the distribution matrix

distr.matrix <- matrix(c(NA, "pois", "geom", "nbinom", 
                         "geom", NA, "pois", "dweibull",
                         "pois", "pois", NA, "geom", 
                         "pois", "geom", "geom", NA), 
                       nrow = s, ncol = s, byrow = TRUE)

# Creation of an array containing the parameters
param1.matrix <- matrix(c(NA, 2, 0.4, 4, 
                          0.7, NA, 5, 0.6, 
                          2, 3, NA, 0.6, 
                          4, 0.3, 0.4, NA), 
                        nrow = s, ncol = s, byrow = TRUE)

param2.matrix <- matrix(c(NA, NA, NA, 0.6, 
                          NA, NA, NA, 0.8, 
                          NA, NA, NA, NA, 
                          NA, NA, NA, NA), 
                        nrow = s, ncol = s, byrow = TRUE)

param.array <- array(c(param1.matrix, param2.matrix), c(s, s, 2))

# Specify the semi-Markov model
smm1 <- smmparametric(states = states, init = vect.init, ptrans = pij, 
                      type.sojourn = "fij", distr = distr.matrix, 
                      param = param.array)

seq1 <- simulate(object = smm1, nsim = c(1000, 10000, 2000), seed = 100)
seq1[[1]][1:15]

}
\seealso{
\link{smmnonparametric}, \link{fitsemimarkovmodel}
}
