% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.mediation.grplasso.fold.R
\name{sparse.mediation.grplasso.fold}
\alias{sparse.mediation.grplasso.fold}
\title{sparse.mediation.grplasso.fold}
\usage{
sparse.mediation.grplasso.fold(
  fold,
  Y,
  X,
  M,
  cvid,
  lambda1 = 0.01,
  lambda2 = 1,
  tol = 10^(-10),
  max.iter = 100,
  grpgroup = c(1, rep(1:V + 1, 2)),
  penalty.factor = c(0, rep(1, V)),
  threshold = 1e-05,
  verbose = FALSE
)
}
\description{
sparse.mediation.grplasso.fold
}
\keyword{internal}
