% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.mediation.grplasso.R
\name{sparse.mediation.grplasso}
\alias{sparse.mediation.grplasso}
\title{Conduct sparse mediation with group LASSO}
\usage{
sparse.mediation.grplasso(
  X,
  M,
  Y,
  tol = 10^(-10),
  max.iter = 100,
  lambda1 = 0.01,
  lambda2 = 1,
  alpha = 0,
  group.penalty.factor = c(0, rep(1, ncol(M))),
  penalty.factor = c(0, rep(1, ncol(M))),
  Omega.out = FALSE,
  verbose = FALSE,
  threshold = 0,
  non.zeros.stop = ncol(M)
)
}
\arguments{
\item{X}{One-dimensional predictor}

\item{M}{Multivariate mediator}

\item{Y}{Outcome}

\item{tol}{(default -10^(-10)) convergence criterion}

\item{max.iter}{(default=100) maximum iteration}

\item{lambda1}{(default=exp(-5:0)) tuning parameter for c',a, b coefficients}

\item{lambda2}{(default=exp(seq(0,0.5*log(ncol(M)),length=3))) tuning parameter for the Omega=Sigma_2^{-1}}

\item{alpha}{(default=0) alpha=0, group lasso penalization will run}

\item{group.penalty.factor}{(V+1)-dimensional group penalty factor vector. If a user does not want to penalize mediator, specify 0 otherwise 1. The first element is the direct effect followed by V-mediators. The default value is c(0,rep(1,V)).}

\item{penalty.factor}{(1+2*V)-dimensional sparsity penalty factor vector.}

\item{threshold}{(default=10^(-8))}

\item{non.zeros.stop}{(default=ncol(M)) when to stop searching regularization path}
}
\value{
c directeffect

hatb Path b (M->Y given X) estimates

hata Path a (X->M) estimates

medest Mediation estimates (a*b)

alpha

lambda

nump Number of selected mediation paths
}
\description{
Fit a mediation model via penalized maximum likelihood and structural equation model.
The regularization path is computed using group lasso at a grid of
values for the regularization parameter lambda. Currently, mediation analysis is developed
based on gaussian assumption.
}
\details{
Multiple Mediaton Model:
(1) M = Xa + e1
(2) Y = Xc' + Mb + e2
And in the optimization, we do not regularize c', due to the assumption of partial mediation.
}
\examples{
N=100
V=50
set.seed(1234)
a = c(rep(1,3),rep(0,V-3))*5;b<-a
X = rnorm(N)
M =  X \%*\% t(a)+ matrix(rnorm(N*V),N,V)
Y =  X + M \%*\% b + rnorm(N)
fit=sparse.mediation.grplasso(X,M,Y,verbose=FALSE, lambda1 = exp(-5:0))
}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{glmnet}
\keyword{highdimensional}
\keyword{mediation}
